+++ src/auth-methods/AwsAuth.ts
@@ -1,3 +1,4 @@
+import aws4 from 'aws4'
 import { BaseAuth } from './BaseAuth'
 import { IAwsAuth } from '../interfaces/auth-methods/IAwsAuth'
 import { ISecret } from '../interfaces/IBaseClient'
@@ -35,6 +36,36 @@ export class AwsAuth extends BaseAuth {
     return {
       result: res.body
     }
+  }
+
+  /**
+   * Wrapper around loginUsingIam() that performs the additional steps
+   * necessary to construct a proper IAM login request to get a token
+   */
+  async getTokenUsingIamLogin (
+    payload: IAwsAuth.IGetTokenUsingIamLoginPayload
+  ): Promise<IVaultResponse<ISecret>> {
+    const body = 'Action=GetCallerIdentity&Version=2011-06-15'
+    const stsUrl = payload.stsUrl || 'https://sts.amazonaws.com/'
+
+    const { headers } = aws4.sign(
+      {
+        service: 'sts',
+        body,
+        headers: payload.iamRequestHeaders
+      },
+      payload.credentials
+    )
+
+    return this.loginUsingIam({
+      role: payload.role,
+      iam_http_request_method: 'POST',
+      iam_request_url: Buffer.from(stsUrl).toString('base64'),
+      iam_request_body: Buffer.from(body).toString('base64'),
+      iam_request_headers: Buffer.from(JSON.stringify(headers)).toString(
+        'base64'
+      )
+    })
   }
 
   /**