+++ src/auth-token-helpers/index.ts
@@ -1,0 +1,4 @@
+import { getTokenFromFile } from './get-token-from-file'
+import { getTokenUsingIam } from './get-token-using-iam'
+
+export { getTokenFromFile, getTokenUsingIam }